Update: refactored function structures to make the scripts lightweight and modular
